requirements.txt
parameterized==0.9.0
PyYAML==6.0.2
